Making the GetSignInLink API match service
